Problem: iPlant Collaborative is now called CyVerse
Some of the links are broken after the name change.

Solution: Fixed the links. Also changed most mentions of iplant/iPlant to cyverse/CyVerse where it makes sense, and where the change has happened.

Note: More changes in the future when the Data Store paths and the wiki domain change to using CyVerse instead of the iPlant.